+++ js/js-structure/js-structure_shapes/index.html
@@ -8,7 +8,7 @@
     />
     <title>HTML & CSS & JS</title>
     <link rel="stylesheet" href="./styles/styles.css" />
-    <script src="./index.js" defer></script>
+    <script src="./index.js" type="module" defer></script>
   </head>
   <body id="root"></body>
 </html>
